requirements.txt
# NoteFlow Requirements
# =====================
# 
# Note: This project requires local packages from sibling repositories:
#   pip install -e /path/to/ai_engine              # ai_core package
#   pip install -e /path/to/notion_md_converter    # notion_markdown_converter package
#
# Minimum Python version: 3.10 (required for ai_core type hints)
# Recommended: Python 3.11+

# Core async and file handling
aiofiles>=24.1.0
aiohttp>=3.11.0
async-timeout>=5.0.0

# Type annotations
typing_extensions>=4.12.0

# AI/ML APIs
anthropic>=0.42.0
openai>=1.59.0
google-ai-generativelanguage>=0.6.10
google-generativeai>=0.8.3
assemblyai>=0.36.0

# HTTP clients
anyio>=4.8.0
httpx>=0.28.0
requests>=2.32.0
urllib3>=2.3.0

# Google APIs (for GDoc integration)
google-api-core>=2.24.0
google-api-python-client>=2.157.0
google-auth>=2.37.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Data validation
pydantic>=2.10.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Media processing
moviepy>=2.1.0
mutagen>=1.47.0

# YAML/Config
PyYAML>=6.0.2

# Environment
python-dotenv>=1.0.1

# Discord
discord.py>=2.0.0

# Scheduling
apscheduler>=3.0.0

# Windows-only packages (will be skipped on macOS/Linux)
pywin32>=308; sys_platform == 'win32'

